+++ languages/go/query.go
@@ -9,7 +9,7 @@ import (
 type Query struct {
 	ffiQuery QueryFfi
 	host     Host
-	calls    map[int]chan interface{}
+	calls    map[uint64]chan interface{}
 }
 
 // NATIVE_TYPES = [int, float, bool, str, dict, type(None), list]
@@ -18,7 +18,7 @@ func newQuery(ffiQuery QueryFfi, host Host) Query {
 	return Query{
 		ffiQuery: ffiQuery,
 		host:     host,
-		calls:    make(map[int]chan interface{}),
+		calls:    make(map[uint64]chan interface{}),
 	}
 }
 
@@ -127,7 +127,7 @@ func (q Query) handleExternalCall(event QueryEventExternalCall) error {
 		method := reflect.ValueOf(instance).MethodByName(string(event.Attribute))
 		if !method.IsValid() {
 			q.ffiQuery.applicationError((&MissingAttributeError{instance: instance, field: string(event.Attribute)}).Error())
-			q.ffiQuery.callResult(int(event.CallId), nil)
+			q.ffiQuery.callResult(event.CallId, nil)
 			return nil
 		}
 		if method.Kind() == reflect.Func {
@@ -206,7 +206,7 @@ func (q Query) handleExternalCall(event QueryEventExternalCall) error {
 		attr := reflect.ValueOf(instance).FieldByName(string(event.Attribute))
 		if !attr.IsValid() {
 			q.ffiQuery.applicationError((&MissingAttributeError{instance: instance, field: string(event.Attribute)}).Error())
-			q.ffiQuery.callResult(int(event.CallId), nil)
+			q.ffiQuery.callResult(event.CallId, nil)
 			return nil
 		}
 		result = attr.Interface()
@@ -216,14 +216,14 @@ func (q Query) handleExternalCall(event QueryEventExternalCall) error {
 	if err != nil {
 		return err
 	}
-	return q.ffiQuery.callResult(int(event.CallId), &Term{*polarValue})
+	return q.ffiQuery.callResult(event.CallId, &Term{*polarValue})
 }
 func (q Query) handleExternalIsa(event QueryEventExternalIsa) error {
 	isa, err := q.host.isa(event.Instance, string(event.ClassTag))
 	if err != nil {
 		return err
 	}
-	return q.ffiQuery.questionResult(int(event.CallId), isa)
+	return q.ffiQuery.questionResult(event.CallId, isa)
 }
 
 func (q Query) handleExternalIsSubSpecializer(event QueryEventExternalIsSubSpecializer) error {
@@ -231,7 +231,7 @@ func (q Query) handleExternalIsSubSpecializer(event QueryEventExternalIsSubSpeci
 	if err != nil {
 		return err
 	}
-	return q.ffiQuery.questionResult(int(event.CallId), res)
+	return q.ffiQuery.questionResult(event.CallId, res)
 }
 
 func (q Query) handleExternalIsSubclass(event QueryEventExternalIsSubclass) error {
@@ -239,7 +239,7 @@ func (q Query) handleExternalIsSubclass(event QueryEventExternalIsSubclass) erro
 	if err != nil {
 		return err
 	}
-	return q.ffiQuery.questionResult(int(event.CallId), res)
+	return q.ffiQuery.questionResult(event.CallId, res)
 }
 
 func (q Query) handleExternalUnify(event QueryEventExternalUnify) error {
@@ -247,7 +247,7 @@ func (q Query) handleExternalUnify(event QueryEventExternalUnify) error {
 	if err != nil {
 		return err
 	}
-	return q.ffiQuery.questionResult(int(event.CallId), res)
+	return q.ffiQuery.questionResult(event.CallId, res)
 }
 
 func (q Query) handleExternalOp(event QueryEventExternalOp) error {
@@ -282,30 +282,30 @@ func (q Query) handleExternalOp(event QueryEventExternalOp) error {
 	default:
 		return fmt.Errorf("Unsupported operation: %v", event.Operator.OperatorVariant)
 	}
-	return q.ffiQuery.questionResult(int(event.CallId), answer)
+	return q.ffiQuery.questionResult(event.CallId, answer)
 }
 
 func (q Query) handleNextExternal(event QueryEventNextExternal) error {
-	if _, ok := q.calls[int(event.CallId)]; !ok {
+	if _, ok := q.calls[event.CallId]; !ok {
 		instance, err := q.host.toGo(event.Iterable)
 		if err != nil {
 			return err
 		}
 		if iter, ok := instance.(Iterator); ok {
-			q.calls[int(event.CallId)] = iter.Iter()
+			q.calls[event.CallId] = iter.Iter()
 		} else {
 			return &InvalidIteratorError{instance: event.Iterable.Value}
 		}
 	}
 
-	iter := q.calls[int(event.CallId)]
+	iter := q.calls[event.CallId]
 	nextValue, ok := <-iter
 	if !ok { // iterator is done
-		return q.ffiQuery.callResult(int(event.CallId), nil)
+		return q.ffiQuery.callResult(event.CallId, nil)
 	}
 	retValue, err := q.host.toPolar(nextValue)
 	if err != nil {
 		return err
 	}
-	return q.ffiQuery.callResult(int(event.CallId), &Term{*retValue})
+	return q.ffiQuery.callResult(event.CallId, &Term{*retValue})
 }
